lab2/pyproject.toml
[project]
name = "lab2"
version = "0.1.0"
description = ""
authors = [
    {name = "Mikestriken",email = "michael.john.marais379@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pytest-asyncio (>=0.25.3,<0.26.0)",
    "aiofiles (>=24.1.0,<25.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
